slab_ptr not supported <5.17
As discussed in issue #1, the introduction of the slab_ptr feature made KRWX not buildable for <=5.16 kernels, since virt_to_folio has been introduced in 5.17. As a temporary workaround, return -ETTY and act accordingly from the user-land library. I plan to use an alternative to virt_to_folio in the future. Also, the same thing applies to the struct local_lock_t, that I fixed.